Move toy location
